Fixed bug that was causing proxychains to break
Proxychains is annoying and wants to resolve DNS names itself; it sends a CONNECT message to the SOCKS server to connect to the system DNS server IP address on 53/tcp. This may be a vestige of SOCKSv4, which did not support remote DNS resolution. The way a SOCKSv5 client normally does remote resolutions is by sending a CONNECT message for a hostname to the remote SOCKS server, which is then forced to resolve the domain itself.

Long story short, I had a type error that I did not catch when an IP address is specified in a CONNECT message that is fixed now.